[Refactor] PP: trim meta comments, hoist imports, consolidate PP/PD tests
Source style pass over the PP/CPP/PD code: drop plan-phase labels from
comments, hoist mid-function imports (pickle, _split_tensor_dict,
get_pp_indices) to module top, remove a stale noqa, and tighten verbose
rationale docstrings to their load-bearing content.

Consolidate tests/pp/ (11 files) + tools/pp/test_custom_collective.py
into tests/test_pp.py and tests/test_pd_pp.py. The new tests exercise
real production functions (Scheduler._prefill_chunk_for_budget,
ScheduledBatch.produces_output, MooncakeConnector._record_write_done,
consumer_region_indices, ...) instead of copied reimplementations, and
drop the stale/duplicate cases. Add CUDAGraphMode / LLMEngine / config
stubs to conftest so the arg-plumbing tests collect GPU-free.
