--- tests/pp/test_pd_pp_port.py
@@ -1,107 +1,0 @@
-# SPDX-License-Identifier: MIT
-# CPP+PD J0 — pp-aware side-channel port offset + remote_pp_size plumbing (GPU-free).
-
-from types import SimpleNamespace
-
-import pytest
-
-from atom.kv_transfer.disaggregation.port_offset import side_channel_port_offset
-from atom.kv_transfer.disaggregation.types import ConnectorMetadata
-
-
-def test_port_offset_pp1_matches_legacy():
-    # pp_rank=0, pp_size=1 must reproduce the old dp_rank*tp_size + tp_rank.
-    for dp_size in (1, 2, 4):
-        for tp_size in (1, 2, 8):
-            for dp_rank in range(dp_size):
-                for tp_rank in range(tp_size):
-                    legacy = dp_rank * tp_size + tp_rank
-                    assert side_channel_port_offset(dp_rank, tp_rank, tp_size) == legacy
-                    assert (
-                        side_channel_port_offset(
-                            dp_rank, tp_rank, tp_size, 0, 1, dp_size
-                        )
-                        == legacy
-                    )
-
-
-def test_port_offset_unique_across_pp_dp_tp():
-    # Every (pp, dp, tp) triple within a topology must map to a distinct port.
-    pp_size, dp_size, tp_size = 4, 2, 2
-    seen = {}
-    for pp_rank in range(pp_size):
-        for dp_rank in range(dp_size):
-            for tp_rank in range(tp_size):
-                off = side_channel_port_offset(
-                    dp_rank, tp_rank, tp_size, pp_rank, pp_size, dp_size
-                )
-                key = (pp_rank, dp_rank, tp_rank)
-                assert off not in seen, f"collision {key} vs {seen.get(off)}"
-                seen[off] = key
-    # Dense packing: offsets fill [0, pp*dp*tp).
-    assert sorted(seen) == list(range(pp_size * dp_size * tp_size))
-
-
-def test_port_offset_pp4_tp1_no_collision():
-    # The debug topology: prefill PP4 x TP1 x DP1 — the case that collided before.
-    offs = [side_channel_port_offset(0, 0, 1, pp_rank, 4, 1) for pp_rank in range(4)]
-    assert offs == [0, 1, 2, 3]
-
-
-def test_build_req_meta_reads_remote_pp_size():
-    meta = ConnectorMetadata._build_req_meta(
-        req_id="r0",
-        local_block_ids=[0, 1],
-        kv_transfer_params={
-            "remote_block_ids": [5, 6],
-            "remote_engine_id": "eng",
-            "remote_host": "10.0.0.1",
-            "remote_port": 41000,
-            "remote_handshake_port": 6301,
-            "tp_size": 1,
-            "remote_pp_size": 4,
-        },
-    )
-    assert meta.remote_pp_size == 4
-
-
-def test_build_req_meta_defaults_pp_size_one():
-    meta = ConnectorMetadata._build_req_meta(
-        req_id="r0",
-        local_block_ids=[0],
-        kv_transfer_params={
-            "remote_block_ids": [5],
-            "remote_host": "h",
-            "remote_handshake_port": 6301,
-            "tp_size": 1,
-        },
-    )
-    assert meta.remote_pp_size == 1
-
-
-def test_producer_stage0_advertises_remote_pp_size():
-    # Needs the worker connector's heavy import chain (torch/aiter); runs on CI/GPU.
-    mc = pytest.importorskip(
-        "atom.kv_transfer.disaggregation.mooncake.mooncake_connector"
-    )
-    sched = object.__new__(mc.MooncakeConnectorScheduler)
-    sched.pp_size = 4
-    sched.tp_size = 1
-    sched.dp_rank = 0
-    sched.engine_id = "eng"
-    sched.host_ip = "10.0.0.1"
-    sched.handshake_port = 40000
-    sched.base_handshake_port = 6301
-    sched.is_producer = True
-
-    seq = SimpleNamespace(
-        output_tokens=[7],
-        spec_token_ids=None,
-        block_table=[1, 2, 3],
-        id=99,
-        per_req_cache_group=-1,
-        kv_transfer_params_output=None,
-    )
-    mc.MooncakeConnectorScheduler.request_finished(sched, seq)
-    assert seq.kv_transfer_params_output["remote_pp_size"] == 4
-    assert seq.kv_transfer_params_output["remote_block_ids"] == [1, 2, 3]
