+++ tests/test_pd_pp.py
@@ -1,0 +1,324 @@
+# SPDX-License-Identifier: MIT
+# PD-disaggregation + pipeline-parallel unit tests (GPU-free).
+
+import threading
+from types import SimpleNamespace
+
+import pytest
+
+from atom.kv_transfer.disaggregation.port_offset import (
+    consumer_region_indices,
+    side_channel_port_offset,
+)
+from atom.kv_transfer.disaggregation.types import ConnectorMetadata
+
+# ---------------------------------------------------------------------------
+# pp-aware side-channel port offset
+# ---------------------------------------------------------------------------
+
+
+def test_port_offset_pp1_matches_legacy():
+    # pp_rank=0, pp_size=1 must reproduce the old dp_rank*tp_size + tp_rank.
+    for dp_size in (1, 2, 4):
+        for tp_size in (1, 2, 8):
+            for dp_rank in range(dp_size):
+                for tp_rank in range(tp_size):
+                    legacy = dp_rank * tp_size + tp_rank
+                    assert side_channel_port_offset(dp_rank, tp_rank, tp_size) == legacy
+                    assert (
+                        side_channel_port_offset(
+                            dp_rank, tp_rank, tp_size, 0, 1, dp_size
+                        )
+                        == legacy
+                    )
+
+
+def test_port_offset_unique_across_pp_dp_tp():
+    pp_size, dp_size, tp_size = 4, 2, 2
+    seen = {}
+    for pp_rank in range(pp_size):
+        for dp_rank in range(dp_size):
+            for tp_rank in range(tp_size):
+                off = side_channel_port_offset(
+                    dp_rank, tp_rank, tp_size, pp_rank, pp_size, dp_size
+                )
+                key = (pp_rank, dp_rank, tp_rank)
+                assert off not in seen, f"collision {key} vs {seen.get(off)}"
+                seen[off] = key
+    # Dense packing: offsets fill [0, pp*dp*tp).
+    assert sorted(seen) == list(range(pp_size * dp_size * tp_size))
+
+
+def test_port_offset_pp4_tp1_no_collision():
+    offs = [side_channel_port_offset(0, 0, 1, pp_rank, 4, 1) for pp_rank in range(4)]
+    assert offs == [0, 1, 2, 3]
+
+
+def test_consumer_targets_every_producer_stage_port():
+    # The ports a consumer computes for stages 0..pp-1 must equal the ports each
+    # producer stage binds, or a stage never receives its write_request.
+    base = 6301
+    args = dict(remote_dp_rank=0, remote_tp_rank=0, remote_tp_size=1, remote_dp_size=1)
+    pp_size = 4
+    consumer_ports = {
+        base
+        + side_channel_port_offset(
+            args["remote_dp_rank"],
+            args["remote_tp_rank"],
+            args["remote_tp_size"],
+            stage,
+            pp_size,
+            args["remote_dp_size"],
+        )
+        for stage in range(pp_size)
+    }
+    assert consumer_ports == {base + i for i in range(pp_size)}
+    assert len(consumer_ports) == pp_size
+
+
+# ---------------------------------------------------------------------------
+# remote_pp_size plumbing
+# ---------------------------------------------------------------------------
+
+
+def test_build_req_meta_reads_remote_pp_size():
+    meta = ConnectorMetadata._build_req_meta(
+        req_id="r0",
+        local_block_ids=[0, 1],
+        kv_transfer_params={
+            "remote_block_ids": [5, 6],
+            "remote_engine_id": "eng",
+            "remote_host": "10.0.0.1",
+            "remote_port": 41000,
+            "remote_handshake_port": 6301,
+            "tp_size": 1,
+            "remote_pp_size": 4,
+        },
+    )
+    assert meta.remote_pp_size == 4
+
+
+def test_build_req_meta_defaults_pp_size_one():
+    meta = ConnectorMetadata._build_req_meta(
+        req_id="r0",
+        local_block_ids=[0],
+        kv_transfer_params={
+            "remote_block_ids": [5],
+            "remote_host": "h",
+            "remote_handshake_port": 6301,
+            "tp_size": 1,
+        },
+    )
+    assert meta.remote_pp_size == 1
+
+
+def test_producer_advertises_remote_pp_size():
+    mc = pytest.importorskip(
+        "atom.kv_transfer.disaggregation.mooncake.mooncake_connector"
+    )
+    sched = object.__new__(mc.MooncakeConnectorScheduler)
+    sched.pp_size = 4
+    sched.tp_size = 1
+    sched.dp_rank = 0
+    sched.engine_id = "eng"
+    sched.host_ip = "10.0.0.1"
+    sched.handshake_port = 40000
+    sched.base_handshake_port = 6301
+    sched.is_producer = True
+
+    seq = SimpleNamespace(
+        output_tokens=[7],
+        spec_token_ids=None,
+        block_table=[1, 2, 3],
+        id=99,
+        per_req_cache_group=-1,
+        kv_transfer_params_output=None,
+    )
+    mc.MooncakeConnectorScheduler.request_finished(sched, seq)
+    assert seq.kv_transfer_params_output["remote_pp_size"] == 4
+    assert seq.kv_transfer_params_output["remote_block_ids"] == [1, 2, 3]
+
+
+# ---------------------------------------------------------------------------
+# Producer per-layer region mapping (consumer_region_indices)
+# ---------------------------------------------------------------------------
+
+
+def _starts(partitions):
+    """Global start layer of each stage, given a per-stage layer-count list."""
+    starts, acc = [], 0
+    for p in partitions:
+        starts.append(acc)
+        acc += p
+    return starts
+
+
+def test_region_map_identity_when_pp1():
+    assert consumer_region_indices(156, 78, 0, 78, 1) == list(range(156))
+
+
+def test_region_map_identity_when_empty():
+    assert consumer_region_indices(0, 0, 5, 78, 4) == []
+
+
+def test_region_map_group_major_single_group():
+    # 1 region/layer, stage of 20 layers @ global start 18 → consumer 18..37.
+    assert consumer_region_indices(20, 20, 18, 78, 4) == list(range(18, 38))
+
+
+def test_region_map_group_major_two_groups_mla():
+    # MLA: 2 groups [kv, index], stage=20 layers @ start 18, N=78.
+    got = consumer_region_indices(40, 20, 18, 78, 4)
+    assert got[:20] == list(range(18, 38))
+    assert got[20:] == list(range(78 + 18, 78 + 38))
+
+
+def test_region_map_undefined_when_not_multiple():
+    assert consumer_region_indices(41, 20, 18, 78, 4) is None
+
+
+def test_region_map_stages_tile_consumer_no_overlap():
+    # GLM-5.2: 78 layers, PP4 partition [18,20,20,20], 2 groups (kv + index).
+    partitions, num_hidden, groups = [18, 20, 20, 20], 78, 2
+    covered = []
+    for start, n_local in zip(_starts(partitions), partitions):
+        covered.extend(
+            consumer_region_indices(n_local * groups, n_local, start, num_hidden, 4)
+        )
+    total = num_hidden * groups
+    assert sorted(covered) == list(range(total))
+    assert len(covered) == len(set(covered))  # no overlap
+
+
+def test_region_map_group_major_beats_naive_offset():
+    # Regression guard: a naive additive offset (start_layer*groups + i) would
+    # misroute group-major layouts. Stage1's index-group region 0 (local idx 20)
+    # must land in the consumer's index group (>=78), not at 36+20=56 (kv group).
+    cmap = consumer_region_indices(40, 20, 18, 78, 4)
+    assert cmap[20] == 78 + 18
+    assert cmap[20] != 56
+
+
+# ---------------------------------------------------------------------------
+# Consumer write-done completion counting + nonce validation
+# ---------------------------------------------------------------------------
+
+
+def _make_connector(**overrides):
+    """Real MooncakeConnector with only the fields _record_write_done touches.
+
+    Bypasses __init__ (RDMA/ZMQ). An empty _release_targets makes the real
+    _send_release a no-op on completion, so the genuine method is exercised.
+    """
+    mc = pytest.importorskip(
+        "atom.kv_transfer.disaggregation.mooncake.mooncake_connector"
+    )
+    conn = object.__new__(mc.MooncakeConnector)
+    conn._completion_lock = threading.Lock()
+    conn._fence_lock = threading.Lock()
+    conn._pending_recv_expected = {}
+    conn._pending_recv_stages = {}
+    conn._pending_recv_nonce = {}
+    conn._pending_recv = set()
+    conn._pending_recv_blocks = {}
+    conn._pending_recv_slots = {}
+    conn._blocks_pending_fence = []
+    conn.done_recving = set()
+    conn._scatter_slot = None
+    conn._release_targets = {}
+    for k, v in overrides.items():
+        setattr(conn, k, v)
+    return conn
+
+
+def test_write_done_correct_nonce_accepted():
+    conn = _make_connector()
+    conn._pending_recv_expected["r1"] = 1
+    conn._pending_recv_nonce["r1"] = 4242
+    assert conn._record_write_done("r1", 0, 0, 4242)
+    assert "r1" in conn.done_recving
+
+
+def test_write_done_wrong_nonce_rejected():
+    conn = _make_connector()
+    conn._pending_recv_expected["r1"] = 1
+    conn._pending_recv_nonce["r1"] = 12345
+    assert not conn._record_write_done("r1", 0, 0, 99999)
+    assert "r1" not in conn.done_recving
+    assert "r1" in conn._pending_recv_expected
+
+
+def test_write_done_zero_nonce_skips_validation():
+    """Old producers that don't send a nonce (default 0) are accepted."""
+    conn = _make_connector()
+    conn._pending_recv_expected["r1"] = 1
+    conn._pending_recv_nonce["r1"] = 0
+    assert conn._record_write_done("r1", 0, 0, 0)
+    assert "r1" in conn.done_recving
+
+
+def test_write_done_missing_nonce_from_old_producer_rejected():
+    """Consumer expects a nonce but the producer sends 0 → rejected."""
+    conn = _make_connector()
+    conn._pending_recv_expected["r1"] = 1
+    conn._pending_recv_nonce["r1"] = 42
+    assert not conn._record_write_done("r1", 0, 0, 0)
+    assert "r1" not in conn.done_recving
+
+
+def test_write_done_nonce_cleaned_up_on_completion():
+    conn = _make_connector()
+    conn._pending_recv_expected["r1"] = 1
+    conn._pending_recv_nonce["r1"] = 777
+    conn._record_write_done("r1", 0, 0, 777)
+    assert "r1" not in conn._pending_recv_nonce
+
+
+def test_write_done_pp_only_dedup():
+    """PP4, TP symmetric: 4 distinct pp_ranks needed to finalize."""
+    conn = _make_connector()
+    conn._pending_recv_expected["r1"] = 4
+    for pp in range(3):
+        assert not conn._record_write_done("r1", pp, 0, 0)
+    assert conn._record_write_done("r1", 3, 0, 0)
+    assert "r1" in conn.done_recving
+
+
+def test_write_done_duplicate_pp_rank_ignored():
+    """Same pp_rank resent (reliability) is not double-counted."""
+    conn = _make_connector()
+    conn._pending_recv_expected["r1"] = 2
+    assert not conn._record_write_done("r1", 0, 0, 0)
+    assert not conn._record_write_done("r1", 0, 0, 0)
+    assert conn._record_write_done("r1", 1, 0, 0)
+
+
+def test_write_done_tp_asymmetric_dedup():
+    """PP2 x TP fan-in 2: 4 distinct (pp, tp) pairs needed."""
+    conn = _make_connector()
+    conn._pending_recv_expected["r1"] = 4
+    assert not conn._record_write_done("r1", 0, 0, 0)
+    assert not conn._record_write_done("r1", 0, 1, 0)
+    assert not conn._record_write_done("r1", 1, 0, 0)
+    assert conn._record_write_done("r1", 1, 1, 0)
+    assert "r1" in conn.done_recving
+
+
+def test_write_done_unknown_request_ignored():
+    conn = _make_connector()
+    assert not conn._record_write_done("unknown", 0, 0, 0)
+
+
+def test_write_done_late_duplicate_after_completion_ignored():
+    conn = _make_connector()
+    conn._pending_recv_expected["r1"] = 1
+    assert conn._record_write_done("r1", 0, 0, 0)
+    assert not conn._record_write_done("r1", 0, 0, 0)
+
+
+def test_write_done_blocks_fenced_on_completion():
+    conn = _make_connector()
+    conn._pending_recv_expected["r1"] = 1
+    conn._pending_recv_blocks["r1"] = [10, 20, 30]
+    conn._record_write_done("r1", 0, 0, 0)
+    assert conn._blocks_pending_fence == [10, 20, 30]
